use chain of responsibility to perform checks
