+++ src/FeedIo/Rule/Atom/LinkNode.php
